Usando doubles em vez de inteiros

+++ src/Main.java
@@ -119,7 +119,7 @@ public class Main {
             return;
         }
 
-        int[][] preMatrix = new int[qtyLines][qtyColumns];
+        double[][] preMatrix = new double[qtyLines][qtyColumns];
 
         for (int i = 0; i < qtyLines; i++) {
             for (int j = 0; j < qtyColumns; j++) {
@@ -150,7 +150,7 @@ public class Main {
         System.out.println("Digite a quantidade de colunas da 1ª matriz: ");
         int qtyColumns = scanner.nextInt();
 
-        int[][] preMatrix1 = new int[qtyLines][qtyColumns];
+        double[][] preMatrix1 = new double[qtyLines][qtyColumns];
 
         System.out.println("1ª Matriz:");
         for (int i = 0; i < qtyLines; i++) {
@@ -162,7 +162,7 @@ public class Main {
 
         System.out.println("-------------------- \n");
 
-        int[][] preMatrix2 = new int[qtyLines][qtyColumns];
+        double[][] preMatrix2 = new double[qtyLines][qtyColumns];
 
         System.out.println("2ª Matriz:");
         for (int i = 0; i < qtyLines; i++) {
@@ -193,7 +193,7 @@ public class Main {
         System.out.println("Digite a quantidade de colunas da matriz: ");
         int qtyColumns = scanner.nextInt();
 
-        int[][] preMatrix = new int[qtyLines][qtyColumns];
+        double[][] preMatrix = new double[qtyLines][qtyColumns];
 
         System.out.println("Matriz:");
         for (int i = 0; i < qtyLines; i++) {
@@ -229,7 +229,7 @@ public class Main {
         System.out.println("Digite a quantidade de colunas da 1ª matriz: ");
         int qtyColumns1 = scanner.nextInt();
 
-        int[][] preMatrix1 = new int[qtyLines1][qtyColumns1];
+        double[][] preMatrix1 = new double[qtyLines1][qtyColumns1];
 
         System.out.println("1ª Matriz:");
         for (int i = 0; i < qtyLines1; i++) {
@@ -246,7 +246,7 @@ public class Main {
         System.out.println("Digite a quantidade de colunas da 2ª matriz: ");
         int qtyColumns2 = scanner.nextInt();
 
-        int[][] preMatrix2 = new int[qtyLines2][qtyColumns2];
+        double[][] preMatrix2 = new double[qtyLines2][qtyColumns2];
 
         System.out.println("2ª Matriz:");
         for (int i = 0; i < qtyLines2; i++) {
@@ -279,7 +279,7 @@ public class Main {
         System.out.println("Digite a quantidade de colunas da matriz: ");
         int qtyColumns = scanner.nextInt();
 
-        int[][] preMatrix = new int[qtyLines][qtyColumns];
+        double[][] preMatrix = new double[qtyLines][qtyColumns];
 
         System.out.println("Matriz:");
         for (int i = 0; i < qtyLines; i++) {
@@ -311,7 +311,7 @@ public class Main {
         System.out.println("Digite a quantidade de colunas da matriz: ");
         int qtyColumns = scanner.nextInt();
 
-        int[][] preMatrix = new int[qtyLines][qtyColumns];
+        double[][] preMatrix = new double[qtyLines][qtyColumns];
 
         System.out.println("Matriz:");
         for (int i = 0; i < qtyLines; i++) {
